moving easy_admin configs to config dir
