+++ config/packages/easy_admin/design.yaml
@@ -1,0 +1,60 @@
+easy_admin:
+    design:
+        brand_color: '#00A8C6'
+        assets:
+            # https://symfony.com/doc/master/bundles/EasyAdminBundle/book/design-configuration.html#loading-the-entire-bootstrap-framework
+            css:
+                - 'bundles/easyadmin/bootstrap-all.css'
+
+        menu:
+            -
+                label: 'Trainings'
+                icon: 'graduation-cap'
+                entity: 'Training'
+            -
+                label: 'Terms'
+                icon: 'calendar'
+                entity: 'TrainingTerm'
+            -
+                label: 'New Registrations'
+                icon: 'id-badge'
+                entity: 'NewRegistration'
+            -
+                icon: 'id-badge'
+                entity: 'Registration'
+            -
+                label: 'Trainers'
+                icon: 'train'
+                entity: 'Trainer'
+            -
+                label: 'Marketing'
+                icon: 'mail-bulk'
+                entity: 'MarketingEvent'
+            -
+                icon: 'comment'
+                label: 'Feedback to Revise'
+                entity: 'NewFeedback'
+            -
+                icon: 'comment'
+                entity: 'Feedback'
+            -
+                icon: 'dog'
+                entity: 'Watchdog'
+
+            -
+                label: 'Money'
+                icon: 'money'
+                children:
+                    - { label: 'Provisions', entity: 'TrainingTermProvision' }
+                    - { label: 'Expenses', entity: 'Expense' }
+
+            -
+                label: 'Helpers'
+                icon: 'plus'
+                children:
+                    - { label: 'Generate Certificate', route: 'create_certificate' }
+
+            -
+                label: 'Back to Website'
+                icon: 'arrow-left'
+                route: 'trainings'
