+++ config/packages/easy_admin/entities/expense.yaml
