+++ config/packages/easy_admin/entities/feedback.yaml
@@ -1,0 +1,24 @@
+easy_admin:
+    entities:
+        NewFeedback:
+            class: 'Pehapkari\Training\Entity\TrainingFeedback'
+            list:
+                # see https://symfony.com/doc/master/bundles/EasyAdminBundle/book/list-search-show-configuration.html#static-filters-dql-filter-option
+                dql_filter: 'entity.isRevised = FALSE'
+                sort: 'createdAt'
+                fields: &feedbackFields
+                    - 'isRevised'
+                    - 'name'
+                    - 'text'
+                    - { property: 'text', template: 'snippet/admin/full_text.twig' }
+                    - { property: 'ratingContent', label: 'Rating' }
+                    - 'training'
+                    - { property: 'isShownOnMainPage', label: 'Main Page' }
+                    - 'isPublic'
+
+        Feedback:
+            class: 'Pehapkari\Training\Entity\TrainingFeedback'
+            list:
+                dql_filter: 'entity.isRevised = TRUE'
+                fields: *feedbackFields
+                sort: 'ratingContent'
