+++ config/packages/easy_admin/entities/registration.yaml
@@ -2,6 +2,7 @@ easy_admin:
     entities:
         NewRegistration:
             class: 'Pehapkari\Registration\Entity\TrainingRegistration'
+            controller: 'Pehapkari\Registration\Admin\Controller\AdminRegistrationController'
             list:
                 dql_filter: 'entity.isPaid = FALSE'
                 batch_actions: ['certificate', 'createInvoices']
