+++ config/packages/easy_admin/entities/training_term_provision.yaml
@@ -1,0 +1,14 @@
+easy_admin:
+    entities:
+        TrainingTermProvision:
+            class: 'Pehapkari\Training\Entity\TrainingTerm'
+            controller: 'Pehapkari\Training\Admin\Controller\AdminTrainingTermController'
+            list:
+                sort: ['startDateTime', 'DESC']
+                fields:
+                    - 'training'
+                    - 'participantCount'
+                    - 'startDateTime'
+                    - 'isProvisionPaid'
+                actions:
+                    - { name: 'training_term_provision', type: 'route', label: 'Provision', css_class: 'btn btn-sm btn-warning' }
