+++ config/packages/easy_admin/entities/user.yaml
