+++ config/routes/easy_admin.yaml
@@ -1,4 +1,4 @@
 easy_admin_bundle:
     resource: '@EasyAdminBundle/Controller/AdminController.php'
-    prefix: "/admin"
     type: annotation
+    prefix: "/admin"
