+++ config/services.yaml
@@ -38,5 +38,11 @@ services:
     Symplify\Statie\Generator\RelatedItemsResolver:
         factory: ['@Pehapkari\DependencyInjection\StatieFactory', 'createRelatedItemsResolver']
 
+    Symplify\PackageBuilder\Console\Style\SymfonyStyleFactory: ~
+    Symfony\Component\Console\Style\SymfonyStyle:
+        factory: ['@Symplify\PackageBuilder\Console\Style\SymfonyStyleFactory', 'create']
+
+    Symplify\PackageBuilder\Reflection\PrivatesAccessor: ~
+
     # latte filters in twig
     Latte\Runtime\FilterExecutor: ~
