--- packages/BetterEasyAdmin/config/config.yaml
@@ -1,2 +1,0 @@
-imports:
-    - { resource: "services.yaml" }
