--- packages/KnowHow/templates/know-how/organize_training.twig
@@ -1,27 +1,0 @@
-{% extends 'base.html.twig' %}
-
-{% block main %}
-    <h1>Organizace školení - checklist</h1>
-
-    {# @todo ask Marťa doplnění #}
-
-    <h2>Před školením</h2>
-    <ul>
-        <li>todo</li>
-    </ul>
-
-    <h2>Po vyhlášení školení</h2>
-    <ul>
-        <li>todo</li>
-    </ul>
-
-    <h2>Den školení</h2>
-    <ul>
-        <li>todo</li>
-    </ul>
-
-    <h2>Po školení</h2>
-    <ul>
-        <li>todo</li>
-    </ul>
-{% endblock %}
