--- packages/Mailer/config/services.yaml
@@ -1,7 +1,0 @@
-services:
-    _defaults:
-        autowire: true
-        public: true
-
-    Pehapkari\Mailer\:
-        resource: '../src/'
