--- packages/Training/config/packages/training_easy_admin.yaml
@@ -1,80 +1,0 @@
-easy_admin:
-    entities:
-        Training:
-            class: 'Pehapkari\Training\Entity\Training'
-            list:
-                fields:
-                    - 'name'
-                    - 'duration'
-                    - 'price'
-                    - 'isPublic'
-            form:
-                fields:
-                    - 'name'
-                    - { property: 'certificateFormattedName', type_options: { attr: { rows: '2' } } }
-                    - 'perex'
-                    - { property: 'description', label: 'Description (use Markdown)', type_options: { attr: { rows: '25' } } }
-                    # TIP: custom template - with "{{ value }}" within
-                    # - { property: 'ibu', template: 'EasyAdmin/Beer/fields/ibu.html.twig' }
-
-        TrainingTerm:
-            label: 'Training Term'
-            class: 'Pehapkari\Training\Entity\TrainingTerm'
-            controller: 'Pehapkari\Training\Admin\Controller\AdminTrainingTermController'
-            list:
-                sort: ['startDateTime', 'DESC']
-                batch_actions: ['generateMarketingEvents']
-                # show only those without paid provision
-                dql_filter: 'entity.isProvisionPaid = FALSE'
-                fields:
-                    - 'training'
-                    - 'startDateTime'
-                    - 'registrations'
-
-        TrainingTermProvision:
-            class: 'Pehapkari\Training\Entity\TrainingTerm'
-            controller: 'Pehapkari\Training\Admin\Controller\AdminTrainingTermController'
-            list:
-                sort: ['startDateTime', 'DESC']
-                fields:
-                    - 'training'
-                    - 'participantCount'
-                    - 'startDateTime'
-                    - 'isProvisionPaid'
-                actions:
-                    - { name: 'training_term_provision', type: 'route', label: 'Provision', css_class: 'btn btn-sm btn-warning' }
-
-        Trainer:
-            class: 'Pehapkari\Training\Entity\Trainer'
-
-        NewFeedback:
-            class: 'Pehapkari\Training\Entity\TrainingFeedback'
-            list:
-                # see https://symfony.com/doc/master/bundles/EasyAdminBundle/book/list-search-show-configuration.html#static-filters-dql-filter-option
-                dql_filter: 'entity.isRevised = FALSE'
-                sort: 'createdAt'
-                fields: &feedbackFields
-                    - 'isRevised'
-                    - 'name'
-                    - 'text'
-                    - { property: 'text', template: 'snippet/admin/full_text.twig' }
-                    - { property: 'ratingContent', label: 'Rating' }
-                    - 'training'
-                    - { property: 'isShownOnMainPage', label: 'Main Page' }
-                    - 'isPublic'
-
-        Feedback:
-            class: 'Pehapkari\Training\Entity\TrainingFeedback'
-            list:
-                dql_filter: 'entity.isRevised = TRUE'
-                fields: *feedbackFields
-                sort: 'ratingContent'
-
-        Watchdog:
-            class: 'Pehapkari\Training\Entity\Watchdog'
-            list:
-                fields:
-                    - 'email'
-                    - 'training'
-                    - 'createdAt'
-                    - 'isInformed'
