+++ packages/User/config/config.yaml
@@ -1,3 +1,9 @@
-imports:
-    - { resource: "services.yaml" }
-    - { resource: "easy_admin.yaml" }
+services:
+    _defaults:
+        public: true
+        autowire: true
+
+    Pehapkari\User\:
+        resource: '../src'
+
+    Symfony\Component\EventDispatcher\EventDispatcher: ~
