--- packages/User/config/services.yaml
@@ -1,12 +1,0 @@
-services:
-    _defaults:
-        public: true
-        autowire: true
-
-    Pehapkari\User\:
-        resource: '../src'
-
-    # consider SecurityUser entity decoupling in the future
-    # https://medium.com/@simshaun/decoupling-your-application-user-from-symfonys-security-user-60fa31b4f7f2
-
-    Symfony\Component\EventDispatcher\EventDispatcher: ~
